Migrate density matrix gate tests to tests_new
